Mensajes de alerta

+++ src/app/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from 'src/app/common/product';
 import { ProductService } from 'src/app/services/product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -27,9 +28,31 @@ export class ProductListComponent implements OnInit{
   }
 
   deleteProductById(id:number){
-    this.productService.deleteProductById(id).subscribe(
-      ()=> this.listProducts()
-    );
+
+    Swal.fire({
+      title: 'Está seguro que quiere eliminar el registro?',
+      text: "",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productService.deleteProductById(id).subscribe(
+          ()=> this.listProducts()
+        );
+        Swal.fire(
+          'Productos',
+          'Producto eliminado correctamente.',
+          'success'
+        )
+      }
+    })
+    
+   
+
   }
